Add files via upload
Polynomial Regression is an extension of linear regression that models the relationship between the dependent variable (salary) and an independent variable (such as years of experience) as an nth-degree polynomial. Unlike simple linear regression which fits a straight line, polynomial regression can fit curves to better capture complex, non-linear patterns in the data.